refactor(api): split matrix.js — extract /matrix/data handler (Q1 part 3)
Stacked on part 2. Final slice of the Q1 god-module split: the 680-line
/matrix/data handler (flat grid, roll-up, context-layered, attribute-fold,
inherited-access modes) moves verbatim to routes/matrix/data.js, mounted via
router.use() — public path unchanged. Pruned imports on both sides so each file
imports only what it uses.

matrix.js: 1,679 (pre-split) -> 202 lines. The module is now a thin composition
root (mounts savedFilters/scope/data sub-routers + the small preview, columns,
hierarchy-paths, inheritance-path handlers).

Adds data.test.js — a DB-free smoke test (USE_SQL unset hits the empty-default
branch) covering module load, route registration, and the no-SQL path.

No query logic, routes, shapes, or ordering changed.
